MVP-28 booking-detail-screen 예약 취소사유 바텀시트 적용, cancel-reservation-screen 삭제
