+++ app/screens/index.ts
@@ -56,7 +56,6 @@ export * from "./cg-mypage/cg-mypage-screen"
 export * from "./ye-beom-test/ye-beom-test-screen"
 export * from "./login/login-screen"
 export * from "./write-review/write-review-screen"
-export * from "./cancel-reservation/cancel-reservation-screen"
 export * from "./make-booking/make-booking-screen"
 export * from "./payment/payment-screen"
 export * from "./test-iamport/test-iamport-screen"
